requirements.txt
lark-parser
graphviz